+++ install_packages/vtktud.py
@@ -73,4 +73,6 @@ class VTKTUD(InstallPackage):
     def clean_build(self):
         # nuke the build dir, the source dir is pristine and there is
         # no installation
-        shutil.rmtree(self.build_dir)
+        utils.output("Removing build dir.")
+        if os.path.exist(self.build_dir):
+            shutil.rmtree(self.build_dir)
